+++ create_fcc.f90
@@ -9,11 +9,11 @@
       SUBROUTINE create_fcc()
       use configs
       IMPLICIT NONE
-      integer :: islx
-      integer :: isly,islz
+      integer(KIND=8) :: islx
+      integer(KIND=8) :: isly,islz
       
-      INTEGER, allocatable :: id(:)
-      integer :: ijk
+      INTEGER(KIND=8), allocatable :: id(:)
+      integer(KIND=8) :: ijk
       REAL (KIND=DBL) :: ww(3)
 
       islx=box_nunits
@@ -78,12 +78,12 @@ END SUBROUTINE create_fcc
       SUBROUTINE define_atom_position_fcc(nn,xx)
       use configs
       IMPLICIT NONE
-      integer, INTENT(in) :: nn ! atom number
-      integer :: isly,islz,islx,mm
+      integer(KIND=8), INTENT(in) :: nn ! atom number
+      integer(KIND=8) :: isly,islz,islx,mm
       REAL (KIND=DBL), intent(out) :: xx(3) 
       
 
-      integer :: ijk
+      integer(KIND=8) :: ijk
 
       islx=box_nunits
       isly=islx
